touch .enabled to enable the scriptworker
We found that on reimage, we start up the scriptworker. This can lead to
bustage because the scriptworker will start without any signing secrets
on disk.

I thought about moving the plist elsewhere, but then the logic to
restart the scriptworker on change will get a lot trickier. Moving the
launchctl commands into a wrapper that check for an .enabled file and
can both start the service for the first time or restart it on a
subsequent change seems to be a good solution.