don't start scriptworker or poller unless .enabled is touched
I thought that we could control this by not running `launchctl load`
after puppetizing. However, the initial image+puppet run reboots, and
MacOS loads all the .plist files after that reboot, so we end up loading
the .plist and starting up scriptworker. By adding the check in the
wrapper scripts as well, we should avoid this.