Don't loop forever in run-puppet.sh on mac
When using the puppet::periodic class we can end up in a situation where
run-puppet.sh is buggy and instead of fixing itself on a later periodic
run when it pulls a fix from git, it just keeps failing until it's
manually killed or the host gets rebooted.

So limit the loop to 5 iterations, then give up.